+++ scanpipe/tests/pipes/test_spdx.py
@@ -44,7 +44,7 @@ class ScanPipeSPDXPipesTest(TestCase):
                 "Organization: Starship ()",
                 "Tool: SPDXCode-1.0",
             ],
-            "licenseListVersion": "3.18",
+            "licenseListVersion": "3.20",
             "comment": "Generated with SPDXCode",
         }
         self.checksum_sha1_data = {
@@ -194,7 +194,7 @@ class ScanPipeSPDXPipesTest(TestCase):
                     "Organization: Starship ()",
                     "Tool: SPDXCode-1.0",
                 ],
-                "licenseListVersion": "3.18",
+                "licenseListVersion": "3.20",
                 "comment": "Generated with SPDXCode",
             },
             "packages": [