+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ Changelog
 v33.0.0 (unreleased)
 --------------------
 
+- Update the CycloneDX BOM download file extension from ``.bom.json`` to ``.cdx.json``.
+  https://github.com/nexB/scancode.io/issues/785
+
 v32.3.0 (2023-06-12)
 --------------------